requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
django-blacklist==0.6.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
django-security==0.12.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.2
openpyxl==3.0.7
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.0.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.0
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
requests==2.26.0
response==0.5.0
scipy==1.7.1
Send2Trash==1.7.1
six==1.16.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
ua-parser==0.10.0
urllib3==1.26.6
virtualenv==20.6.0
virtualenv-clone==0.5.5
wcwidth==0.2.5
webencodings==0.5.1
